Moved LoggedInDataAccessInterface to data_access package
