+++ src/main/java/use_case/login_user_story/login/LoginInteractor.java
@@ -2,7 +2,7 @@ package use_case.login_user_story.login;
 
 import data_access.LoggedInDataAccessObject;
 import entity.User;
-import use_case.user_profile_user_story.user_profile.LoggedInDataAccessInterface;
+import data_access.LoggedInDataAccessInterface;
 
 /**
  * The Login Interactor.
